Satisfy viewer structure guard
